requirements.txt
# Trading Bot Dependencies

# Market Data
yfinance>=0.2.66

# Data Processing
pandas>=1.3.0
numpy>=1.16.5

# Alpaca Trading (MCP)
# No pip install needed - uses Cursor MCP integration

# Knowledge Graph (MCP)
# No pip install needed - uses Cursor MCP integration

# SATBASE System-1
httpx>=0.27.0
pydantic>=2.7.0
polars>=0.20.0
pyarrow>=15.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
tenacity>=8.2.3
python-dotenv>=1.0.1
PyYAML>=6.0.2
python-dateutil>=2.9.0.post0
feedparser>=6.0.10
